Support "Reboot to Recovery" by using a script to edit cmdline
Use a script to edit androidboot.mode=, so that we can be able to switch
between recovery or system. Use sys.shutdown.requested to trigger the script
and edit the same file that bootctrl HAL uses then reboot.

For recovery, remember to port sys.shutdown.requested by using this commit:
https://github.com/BlissOS/platform_bootable_recovery/commit/abe6110007f658acc823e4070d38134aa1ccc55c

Big thanks to 0xCAFEBABE for the idea

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>